app/requirements.txt
fastapi==0.92.0
pydantic==1.10.5
uvicorn==0.20.0
passlib==1.7.4
python-jose==3.3.0
mysql-connector-python==8.0.32